inventory-service/requirements.txt
Flask==2.3.3
kafka-python==2.0.2
python-dotenv==1.0.0
prometheus-client==0.17.1
